JACOBIN-603 reorg gfunctions; no functional changes except <clinit> returns an identifying string
