+++ src/gfunction/javaIoFile.go
@@ -21,13 +21,13 @@ func Load_Io_File() {
 	MethodSignatures["java/io/File.<clinit>()V"] =
 		GMeth{
 			ParamSlots: 0,
-			GFunction:  justReturn,
+			GFunction:  clinitGeneric,
 		}
 
 	MethodSignatures["java/io/File.<clinit>()V"] =
 		GMeth{
 			ParamSlots: 0,
-			GFunction:  justReturn,
+			GFunction:  clinitGeneric,
 		}
 
 	MethodSignatures["java/io/File.<init>(Ljava/lang/String;)V"] =
