+++ src/gfunction/javaLangByte.go
@@ -20,7 +20,7 @@ func Load_Lang_Byte() {
 	MethodSignatures["java/lang/Byte.<clinit>()V"] =
 		GMeth{
 			ParamSlots: 0,
-			GFunction:  justReturn,
+			GFunction:  clinitGeneric,
 		}
 
 	MethodSignatures["java/lang/Byte.decode(Ljava/lang/String;)Ljava/lang/Byte;"] =
