+++ src/gfunction/javaLangDouble.go
@@ -22,7 +22,7 @@ func Load_Lang_Double() {
 	MethodSignatures["java/lang/Double.<clinit>()V"] =
 		GMeth{
 			ParamSlots: 0,
-			GFunction:  justReturn,
+			GFunction:  clinitGeneric,
 		}
 
 	MethodSignatures["java/lang/Double.byteValue()B"] =
