+++ src/gfunction/javaLangLong.go
@@ -20,7 +20,7 @@ func Load_Lang_Long() {
 	MethodSignatures["java/lang/Long.<clinit>()V"] =
 		GMeth{
 			ParamSlots: 0,
-			GFunction:  justReturn,
+			GFunction:  clinitGeneric,
 		}
 
 	MethodSignatures["java/lang/Long.doubleValue()D"] =
